// src/app/models/simulationResult.ts
import { EconomicEngineResponse } from './economicEngine';

/**
 * DSF-SE API
 * Sample version of DSF-SE API for communication with the professional UI
 *
 * OpenAPI spec version: 3
 * Contact: sisay.adugna.chala@fit.fraunhofer.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ResponseMessage {
  from: string;
  simulationID?: string;
  info?: string;
  status?: number;
  result?: SimulationResult;
}

export interface Phases {
  'Phase 1': Phase;
  'Phase 2'?: Phase;
  'Phase 3'?: Phase;
}

export interface Currents {
  [key: string]: Phases;
}

export interface Losses {
  circuit_total_losses: number;
  Line: { [key: string]: number };
  [key: string]: { [key: string]: number } | number;
}

export interface Powers {
  [key: string]: { [key: string]: number };
}

export interface Phase {
  max: number;
  min: number;
}

export interface Voltages {
  [key: string]: Phases;
}

export interface SimulationResult {
  currents: Currents;
  losses: Losses;
  powers: Powers;
  voltages: Voltages;
  economicEngine: EconomicEngineResponse;
}